requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask_cors==5.0.0
waitress==3.0.0
requests==2.32.3
matplotlib==3.9.2
pillow==10.4.0